requirements.txt
# Runtime dependencies — pinned loose on purpose so a fresh clone
# pulls compatible up-to-date versions.  CI should pin-lock this.

# --- Core ---
pydantic>=2.0,<3.0
pyyaml>=6.0
httpx>=0.27
numpy>=1.24,<3.0

# --- LLM client (Vertex AI via google-genai) ---
google-genai>=0.5
google-auth>=2.30

# --- Semantic search ---
faiss-cpu>=1.10                  # CPU FAISS is fine for runtime queries

# --- Data-layer hydration (HF artifact downloads) ---
huggingface_hub>=0.24            # snapshot_download for /data layer

# --- Dev / test ---
pytest>=8.0
pytest-asyncio>=0.23
tqdm>=4.66
